+++ spec/www/spec/legacy.js
@@ -52,7 +52,7 @@ var mytests = function() {
 
   for (var i=0; i<scenarioCount; ++i) {
 
-    describe(scenarioList[i] + ': legacy sql test(s)', function() {
+    describe(scenarioList[i] + ': LEGACY SQL test(s)', function() {
       var scenarioName = scenarioList[i];
       var suiteName = scenarioName + ': ';
       var isWebSql = (i === 1);