+++ spec/www/spec/simple-test.js
@@ -4,11 +4,6 @@ var MYTIMEOUT = 12000;
 
 var DEFAULT_SIZE = 5000000; // max to avoid popup in safari/ios
 
-// FUTURE TODO replace in test(s):
-function ok(test, desc) { expect(test).toBe(true); }
-function equal(a, b, desc) { expect(a).toEqual(b); } // '=='
-function strictEqual(a, b, desc) { expect(a).toBe(b); } // '==='
-
 var isAndroid = /Android/.test(navigator.userAgent);
 var isWP8 = /IEMobile/.test(navigator.userAgent); // Matches WP(7/8/8.1)
 //var isWindows = /Windows NT/.test(navigator.userAgent); // Windows [NT] (8.1)
@@ -31,7 +26,7 @@ var mytests = function() {
 
   for (var i=0; i<scenarioCount; ++i) {
 
-    describe(scenarioList[i] + ': simple test(s)', function() {
+    describe(scenarioList[i] + ': SIMPLE SQL test(s)', function() {
       var scenarioName = scenarioList[i];
       var suiteName = scenarioName + ': ';
       var isWebSql = (i === 1);
@@ -43,25 +38,25 @@ var mytests = function() {
           return window.sqlitePlugin.openDatabase({name: name, androidDatabaseImplementation: 2});
         }
         if (isWebSql) {
-          return window.openDatabase(name, "1.0", "Demo", DEFAULT_SIZE);
+          return window.openDatabase(name, '1.0', 'Test', DEFAULT_SIZE);
         } else {
-          return window.sqlitePlugin.openDatabase(name, "1.0", "Demo", DEFAULT_SIZE);
+          return window.sqlitePlugin.openDatabase(name, '1.0', 'Test', DEFAULT_SIZE);
         }
       }
 
-      it(suiteName + "US-ASCII String manipulation test",
+      it(suiteName + 'US-ASCII String manipulation test',
         function(done) {
-          var db = openDatabase("ASCII-string-test.db", "1.0", "Demo", DEFAULT_SIZE);
+          var db = openDatabase('ASCII-string-test.db', '1.0', 'Test', DEFAULT_SIZE);
 
-          expect(db).toBeDefined()
+          expect(db).toBeDefined();
 
           db.transaction(function(tx) {
 
-            expect(tx).toBeDefined()
+            expect(tx).toBeDefined();
 
-            tx.executeSql("select upper('Some US-ASCII text') as uppertext", [], function(tx, res) {
-              console.log("res.rows.item(0).uppertext: " + res.rows.item(0).uppertext);
-              expect(res.rows.item(0).uppertext).toEqual("SOME US-ASCII TEXT");
+            tx.executeSql("SELECT UPPER('Some US-ASCII text') AS uppertext", [], function(tx, res) {
+              console.log('res.rows.item(0).uppertext: ' + res.rows.item(0).uppertext);
+              expect(res.rows.item(0).uppertext).toEqual('SOME US-ASCII TEXT');
 
               done();
             });
@@ -70,21 +65,21 @@ var mytests = function() {
 
       // Only test ICU-UNICODE with Android 5.0(+) (Web SQL):
       if (isWebSql && /Android [5-9]/.test(navigator.userAgent))
-        it(suiteName + "ICU-UNICODE string manipulation test", function(done) {
+        it(suiteName + 'ICU-UNICODE string manipulation test', function(done) {
           if ((!isWebSql) && isAndroid) pending('BROKEN for Android version of plugin [with sqlite-connector]');
 
-          var db = openDatabase("UNICODE-string-test.db", "1.0", "Demo", DEFAULT_SIZE);
+          var db = openDatabase('UNICODE-string-test.db', '1.0', 'Test', DEFAULT_SIZE);
 
-          expect(db).toBeDefined()
+          expect(db).toBeDefined();
 
           db.transaction(function(tx) {
 
-            expect(tx).toBeDefined()
+            expect(tx).toBeDefined();
 
             // 'Some Cyrillic text'
-            tx.executeSql("select UPPER('Какой-то кириллический текст') as uppertext", [], function (tx, res) {
-              console.log("res.rows.item(0).uppertext: " + res.rows.item(0).uppertext);
-              expect(res.rows.item(0).uppertext).toEqual("КАКОЙ-ТО КИРИЛЛИЧЕСКИЙ ТЕКСТ");
+            tx.executeSql("SELECT UPPER('Какой-то кириллический текст') AS uppertext", [], function (tx, res) {
+              console.log('res.rows.item(0).uppertext: ' + res.rows.item(0).uppertext);
+              expect(res.rows.item(0).uppertext).toEqual('КАКОЙ-ТО КИРИЛЛИЧЕСКИЙ ТЕКСТ');
 
               done();
             });
@@ -93,22 +88,20 @@ var mytests = function() {
 
         it(suiteName + 'Simple INSERT test: check insertId & rowsAffected in result', function(done) {
 
-          var db = openDatabase("INSERT-test.db", "1.0", "Demo", DEFAULT_SIZE);
+          var db = openDatabase('INSERT-test.db', '1.0', 'Test', DEFAULT_SIZE);
 
-          ok(!!db, "db object");
+          expect(db).toBeDefined();
 
           db.transaction(function(tx) {
-            ok(!!tx, "tx object");
+            expect(tx).toBeDefined();
 
             tx.executeSql('DROP TABLE IF EXISTS test_table');
             tx.executeSql('CREATE TABLE IF NOT EXISTS test_table (id integer primary key, data text, data_num integer)');
 
-            tx.executeSql("INSERT INTO test_table (data, data_num) VALUES (?,?)", ["test", 100], function(tx, res) {
-              console.log("insertId: " + res.insertId + " -- probably 1");
-              console.log("rowsAffected: " + res.rowsAffected + " -- should be 1");
-
-              ok(!!res.insertId, "Valid res.insertId");
-              equal(res.rowsAffected, 1, "res rows affected");
+            tx.executeSql('INSERT INTO test_table (data, data_num) VALUES (?,?)', ['test', 100], function(tx, res) {
+              expect(res).toBeDefined();
+              expect(res.insertId).toBeDefined();
+              expect(res.rowsAffected).toBe(1);
 
               done();
             });
@@ -116,121 +109,120 @@ var mytests = function() {
           });
         }, MYTIMEOUT);
 
-      it(suiteName + "db transaction test",
+      it(suiteName + 'db transaction test',
         function(done) {
-          var db = openDatabase("db-trx-test.db", "1.0", "Demo", DEFAULT_SIZE);
+          var db = openDatabase('db-trx-test.db', '1.0', 'Test', DEFAULT_SIZE);
 
-          ok(!!db, "db object");
+          expect(db).toBeDefined();
 
           var check = 0;
 
           db.transaction(function(tx) {
-
-            ok(!!tx, "tx object");
+            // first tx object:
+            expect(tx).toBeDefined();
 
             tx.executeSql('DROP TABLE IF EXISTS test_table');
             tx.executeSql('CREATE TABLE IF NOT EXISTS test_table (id integer primary key, data text, data_num integer)');
 
-            tx.executeSql("INSERT INTO test_table (data, data_num) VALUES (?,?)", ["test", 100], function(tx, res) {
+            tx.executeSql('INSERT INTO test_table (data, data_num) VALUES (?,?)', ['test', 100], function(tx, res) {
+              // check tx & res object parameters:
               expect(tx).toBeDefined();
               expect(res).toBeDefined();
-
-              console.log("insertId: " + res.insertId + " -- probably 1");
-              console.log("rowsAffected: " + res.rowsAffected + " -- should be 1");
 
               expect(res.insertId).toBeDefined();
               expect(res.rowsAffected).toBe(1);
 
               db.transaction(function(tx) {
-                ok(!!tx, "second tx object");
+                // second tx object:
+                expect(tx).toBeDefined();
 
-                tx.executeSql("SELECT count(id) as cnt from test_table;", [], function(tx, res) {
+                tx.executeSql('SELECT COUNT(id) AS cnt FROM test_table;', [], function(tx, res) {
                   ++check;
 
-                  console.log("res.rows.length: " + res.rows.length + " -- should be 1");
-                  console.log("res.rows.item(0).cnt: " + res.rows.item(0).cnt + " -- should be 1");
-
-                  equal(res.rows.length, 1, "res rows length");
-                  equal(res.rows.item(0).cnt, 1, "select count");
+                  expect(res.rows.length).toBe(1);
+                  expect(res.rows.item(0).cnt).toBe(1);
                 });
 
-                tx.executeSql("SELECT data_num from test_table;", [], function(tx, res) {
+                tx.executeSql('SELECT data_num FROM test_table;', [], function(tx, res) {
                   ++check;
 
-                  equal(res.rows.length, 1, "SELECT res rows length");
-                  equal(res.rows.item(0).data_num, 100, "SELECT data_num");
+                  expect(res.rows.length).toBe(1);
+                  expect(res.rows.item(0).data_num).toBe(100);
                 });
 
-                tx.executeSql("UPDATE test_table SET data_num = ? WHERE data_num = 100", [101], function(tx, res) {
+                tx.executeSql('UPDATE test_table SET data_num = ? WHERE data_num = 100', [101], function(tx, res) {
                   ++check;
-
-                  console.log("UPDATE rowsAffected: " + res.rowsAffected + " -- should be 1");
 
                   expect(res.rowsAffected).toBe(1);
                 });
 
-                tx.executeSql("SELECT data_num from test_table;", [], function(tx, res) {
+                tx.executeSql('SELECT data_num FROM test_table;', [], function(tx, res) {
                   ++check;
 
-                  equal(res.rows.length, 1, "SELECT res rows length");
-                  equal(res.rows.item(0).data_num, 101, "SELECT data_num");
+                  expect(res.rows.length).toBe(1);
+                  expect(res.rows.item(0).data_num).toBe(101);
                 });
 
                 tx.executeSql("DELETE FROM test_table WHERE data LIKE 'tes%'", [], function(tx, res) {
                   ++check;
 
-                  console.log("DELETE rowsAffected: " + res.rowsAffected + " -- should be 1");
-
                   expect(res.rowsAffected).toBe(1);
                 });
 
-                tx.executeSql("SELECT data_num from test_table;", [], function(tx, res) {
+                tx.executeSql('SELECT data_num FROM test_table;', [], function(tx, res) {
                   ++check;
 
-                  equal(res.rows.length, 0, "SELECT res rows length");
+                  expect(res.rows.length).toBe(0);
                 });
 
               }, function(e) {
-                console.log("ERROR: " + e.message);
-                expect(false);
+                // not expected:
+                expect(false).toBe(true);
+                console.log('ERROR: ' + e.message);
               }, function() {
-                console.log("second tx ok success cb");
+                console.log('second tx ok success cb');
                 expect(check).toBe(6);
 
                 done();
               });
 
             }, function(e) {
-              console.log("ERROR: " + e.message);
-              expect(false);
+              // not expected:
+              expect(false).toBe(true);
+              console.log('ERROR: ' + e.message);
             });
           }, function(e) {
-            console.log("ERROR: " + e.message);
-            expect(false);
+            // not expected:
+            expect(false).toBe(true);
+            console.log('ERROR: ' + e.message);
+            done();
           }, function() {
-            console.log("tx success cb");
+            console.log('first tx success cb OK');
           });
 
         }, MYTIMEOUT);
 
-      it(suiteName + "number values inserted using number bindings",
+      it(suiteName + 'number values inserted using number bindings',
         function(done) {
-          var db = openDatabase("Value-binding-test.db", "1.0", "Demo", DEFAULT_SIZE);
+          var db = openDatabase('Value-binding-test.db', '1.0', 'Test', DEFAULT_SIZE);
           db.transaction(function(tx) {
             tx.executeSql('DROP TABLE IF EXISTS test_table');
             tx.executeSql('CREATE TABLE IF NOT EXISTS test_table (id integer primary key, data_text1, data_text2, data_int, data_real)');
-          }, function(err) { ok(false, err.message) }, function() {
+          }, function(err) {
+              // went wrong:
+              expect(false).toBe(true);
+          }, function() {
             db.transaction(function(tx) {
               // create columns with no type affinity
-              tx.executeSql("insert into test_table (data_text1, data_text2, data_int, data_real) VALUES (?,?,?,?)", ["314159", "3.14159", 314159, 3.14159], function(tx, res) {
+              tx.executeSql('INSERT INTO test_table (data_text1, data_text2, data_int, data_real) VALUES (?,?,?,?)', ['314159', '3.14159', 314159, 3.14159], function(tx, res) {
                 expect(res.rowsAffected).toBe(1);
-                tx.executeSql("select * from test_table", [], function(tx, res) {
+                tx.executeSql('SELECT * FROM test_table', [], function(tx, res) {
                   var row = res.rows.item(0);
-                  strictEqual(row.data_text1, "314159", "data_text1 should have inserted data as text");
+                  expect(row.data_text1).toBe('314159'); // data_text1 should have inserted data as text
                   if (!isWP8) // JSON issue in WP(8) version
-                    strictEqual(row.data_text2, "3.14159", "data_text2 should have inserted data as text");
-                  strictEqual(row.data_int, 314159, "data_int should have inserted data as an integer");
-                  ok(Math.abs(row.data_real - 3.14159) < 0.000001, "data_real should have inserted data as a real");
+                    expect(row.data_text2).toBe('3.14159'); // data_text2 should have inserted data as text
+                  expect(row.data_int).toBe(314159); // data_int should have inserted data as an integer
+                  expect(Math.abs(row.data_real - 3.14159) < 0.000001).toBe(true); // data_real should have inserted data as a real
 
                   done();
                 });
@@ -239,11 +231,11 @@ var mytests = function() {
           });
         }, MYTIMEOUT);
 
-        /* thanks to @calebeaires: */
+        /* THANKS to @calebeaires: */
         it(suiteName + 'create virtual table using FTS3', function(done) {
           if (isWP8) pending('NOT IMPLEMENTED for WP(8)'); // NOT IMPLEMENTED in CSharp-SQLite
 
-          var db = openDatabase('virtual-table-using-fts3.db', '1.0', "Demo", DEFAULT_SIZE);
+          var db = openDatabase('virtual-table-using-fts3.db', '1.0', 'Test', DEFAULT_SIZE);
           expect(db).toBeDefined();
 
           db.transaction(function(tx) {
@@ -275,7 +267,7 @@ var mytests = function() {
         it(suiteName + 'create virtual table using FTS4', function(done) {
           if (isWP8) pending('NOT IMPLEMENTED for WP(8)'); // NOT IMPLEMENTED in CSharp-SQLite
 
-          var db = openDatabase('virtual-table-using-fts4.db', '1.0', "Demo", DEFAULT_SIZE);
+          var db = openDatabase('virtual-table-using-fts4.db', '1.0', 'Test', DEFAULT_SIZE);
           expect(db).toBeDefined();
 
           db.transaction(function(tx) {
@@ -307,7 +299,7 @@ var mytests = function() {
           if (isWP8) pending('NOT IMPLEMENTED for WP(8)'); // NOT IMPLEMENTED in CSharp-SQLite
           if (isAndroid) pending('NOT IMPLEMENTED for all versions of Android'); // NOT IMPLEMENTED for all versions of Android database (failed in Circle CI)
 
-          var db = openDatabase('virtual-table-using-r-tree.db', '1.0', "Demo", DEFAULT_SIZE);
+          var db = openDatabase('virtual-table-using-r-tree.db', '1.0', 'Test', DEFAULT_SIZE);
           expect(db).toBeDefined();
 
           db.transaction(function(tx) {
@@ -339,7 +331,7 @@ var mytests = function() {
         it(suiteName + 'Skip callbacks after syntax error with no handler', function(done) {
           if (!isWebSql) pending('Plugin BROKEN'); // XXX TODO
 
-          var db = openDatabase('first-syntax-error-with-no-handler.db', '1.0', "Demo", DEFAULT_SIZE);
+          var db = openDatabase('first-syntax-error-with-no-handler.db', '1.0', 'Test', DEFAULT_SIZE);
           expect(db).toBeDefined();
 
           db.transaction(function(tx) {
@@ -375,7 +367,7 @@ var mytests = function() {
         it(suiteName + 'Skip callbacks after syntax error handler returns true', function(done) {
           if (!isWebSql) pending('Plugin BROKEN'); // XXX TODO
 
-          var db = openDatabase('first-syntax-error-handler-returns-true.db', '1.0', "Demo", DEFAULT_SIZE);
+          var db = openDatabase('first-syntax-error-handler-returns-true.db', '1.0', 'Test', DEFAULT_SIZE);
           expect(db).toBeDefined();
 
           var isFirstErrorHandlerCalled = false; // poor man's spy
@@ -431,7 +423,7 @@ var mytests = function() {
         // according to the spec at http://www.w3.org/TR/webdatabase/ the transaction should be
         // recovered *only* if the sql error handler returns false.
         it(suiteName + 'Recover transaction with callbacks after syntax error handler returns false', function(done) {
-          var db = openDatabase('recover-if-syntax-error-handler-returns-false.db', '1.0', "Demo", DEFAULT_SIZE);
+          var db = openDatabase('recover-if-syntax-error-handler-returns-false.db', '1.0', 'Test', DEFAULT_SIZE);
           expect(db).toBeDefined();
 
           var isFirstErrorHandlerCalled = false; // poor man's spy
@@ -484,7 +476,7 @@ var mytests = function() {
         // NOTE: as discussed in litehelpers/Cordova-sqlite-storage#232 this plugin is correct
         // according to the spec at http://www.w3.org/TR/webdatabase/
         it(suiteName + 'syntax error handler returns undefined', function(done) {
-          var db = openDatabase('syntax-error-handler-returns-undefined.db', '1.0', "Demo", DEFAULT_SIZE);
+          var db = openDatabase('syntax-error-handler-returns-undefined.db', '1.0', 'Test', DEFAULT_SIZE);
           expect(db).toBeDefined();
 
           var isFirstErrorHandlerCalled = false; // poor man's spy